backend/requirements.txt
google-cloud-storage
Flask
Pillow
